refactor(Gradle): Move Git-related code to `ValueSource` providers
This is a better way to make running external processes compatible with
the configuration cache, see [1].

[1]: https://docs.gradle.org/current/userguide/configuration_cache.html#config_cache:requirements:external_processes
